// src/app/model/employee.model.ts
export class Employee {
    code: string;
    name: string;
    gender: string;
    email?: string;
    phoneNumber?: number;
    contactPreference: string;
    dateOfBirth: Date;
    department: string;
    isActive: boolean;
    photoPath?: string;
    // password: string;
    // confirmPassword: string;
}